0.9.53 Performance improvement: create components for frames only when it is necessary.
