+++ content_scripts/content_scripts.css
@@ -72,9 +72,8 @@ div.surfingkeys_cursor:empty {
 }
 #sk_frame {
     position: fixed;
-    border: 4px solid #3CF30F;
+    border: 4px solid #434343;
     box-sizing: border-box;
     z-index: 2147483000;
-    background: rgba(129, 120, 222, 0.62);
-    box-shadow: 0px 0px 10px rgba(218, 60, 13, 0.8);
+    background: rgba(202, 202, 202, 0.62);
 }
