+++ content_scripts/normal.js
@@ -1,4 +1,4 @@
-var Mode = (function() {
+function createMode() {
     var self = function(name, statusLine) {
         this.name = name;
         this.statusLine = statusLine;
@@ -191,7 +191,7 @@ var Mode = (function() {
                     }
                 }
             }
-            Front.showStatus(0, sl);
+            typeof(Front) === "object" && Front.showStatus(0, sl);
         }
     };
 
@@ -284,9 +284,9 @@ var Mode = (function() {
     };
 
     return self;
-})();
+}
 
-var Disabled = (function() {
+function createDisabled() {
     var self = new Mode("Disabled");
 
     // Disabled has higher priority than others.
@@ -303,9 +303,9 @@ var Disabled = (function() {
     });
 
     return self;
-})();
+}
 
-var PassThrough = (function() {
+function createPassThrough() {
     var self = new Mode("PassThrough");
     var _autoExit, _timeout;
 
@@ -344,22 +344,15 @@ var PassThrough = (function() {
     };
 
     return self;
-})();
+}
 
-var Normal = (function() {
+function createNormal() {
     var self = new Mode("Normal");
 
     // let next focus event pass
     var _passFocus = false;
     self.passFocus = function(pf) {
         _passFocus = pf;
-    };
-
-    self.onEnter = function() {
-        if (runtime.conf.stealFocusOnLoad && !Front.isProvider()) {
-            var elm = getRealEdit();
-            elm && elm.blur();
-        }
     };
 
     self.addEventListener('keydown', function(event) {
@@ -416,7 +409,7 @@ var Normal = (function() {
     });
     self.addEventListener('focus', function(event) {
         Mode.showStatus();
-        if (runtime.conf.stealFocusOnLoad && !Front.isProvider()) {
+        if (runtime.conf.stealFocusOnLoad && !isInUIFrame()) {
             var elm = getRealEdit(event);
             if (isEditable(elm)) {
                 if (_passFocus || elm.enableAutoFocus) {
@@ -437,6 +430,10 @@ var Normal = (function() {
         }, 0);
     });
     self.addEventListener('mousedown', function(event) {
+        // Insert mode will never be created in frontend frame.
+        if (typeof(Insert) === "undefined") {
+            return;
+        }
         // The isTrusted read-only property of the Event interface is a boolean
         // that is true when the event was generated by a user action, and false
         // when the event was created or modified by a script or dispatched via dispatchEvent.
@@ -663,7 +660,8 @@ var Normal = (function() {
             scrollNode = scrollNodes[scrollIndex];
             if (scrollNode !== document.scrollingElement) {
                 var br = scrollNode.getBoundingClientRect();
-                if (br.width === 0 || br.height === 0) {
+                if (br.width === 0 || br.height === 0 || !isElementPartiallyInViewport(scrollNode)
+                    || !hasScroll(scrollNode, 'y', 16)) {
                     scrollNodes.splice(scrollIndex, 1);
                     scrollIndex = 0;
                     scrollNode = scrollNodes[scrollIndex];
@@ -733,9 +731,8 @@ var Normal = (function() {
     };
 
     self.addScrollableElement = function(elm) {
-        initScrollIndex();
-        scrollIndex = scrollNodes.indexOf(elm);
-        if (scrollIndex === -1) {
+        if (!scrollNodes || !elm.contains(scrollNodes[scrollIndex]) && scrollNodes.indexOf(elm) === -1) {
+            initScrollIndex();
             scrollNodes.push(elm);
             scrollIndex = scrollNodes.length - 1;
         }
@@ -1140,4 +1137,4 @@ var Normal = (function() {
     });
 
     return self;
-})();
+}
