+++ manifest.json
@@ -2,7 +2,7 @@
     "manifest_version": 2,
     "name": "Surfingkeys",
     "short_name": "Rich shortcuts in vim spirit for productivity with keyboard.",
-    "version": "0.9.52",
+    "version": "0.9.53",
     "description": "Rich shortcuts to click links/switch tabs/scroll pages or capture full page, use Chrome like vim for productivity.",
     "icons": {
         "16": "icons/16.png",
@@ -62,11 +62,10 @@
             ],
             "match_about_blank": true,
             "js": [
-                "content_scripts/common_content.min.js",
+                "content_scripts/modules.min.js",
                 "content_scripts/front.js",
-                "content_scripts/content_scripts.js",
-                "content_scripts/top.js",
-                "pages/default.js"
+                "pages/default.js",
+                "content_scripts/content_scripts.js"
             ],
             "css": [
                 "content_scripts/content_scripts.css"
