+++ pages/error.html
@@ -68,11 +68,10 @@ kbd {
                 </ul>
             </div>
         </div>
-        <script src="../content_scripts/common_content.min.js"></script>
+        <script src="../content_scripts/modules.min.js"></script>
         <script src="../content_scripts/front.js"></script>
-        <script src="../content_scripts/content_scripts.js"></script>
-        <script src="../content_scripts/top.js"></script>
         <script src="default.js"></script>
+        <script src="../content_scripts/content_scripts.js"></script>
         <script src="error.js"></script>
     </body>
 </html>
