+++ pages/frontend.html
@@ -42,7 +42,7 @@
     </div>
     <div id=sk_keystroke class=sk_theme style="display: none;">
     </div>
-    <script src="../content_scripts/common_content.min.js"></script>
+    <script src="../content_scripts/modules.min.js"></script>
     <script src="front.js"></script>
     <script src="omnibar.js"></script>
     <script src="statusbar.js"></script>
