+++ pages/markdown.html
@@ -8,11 +8,10 @@
         <link rel="stylesheet" href="markdown.css">
     </head>
     <body>
-        <script src="../content_scripts/common_content.min.js"></script>
+        <script src="../content_scripts/modules.min.js"></script>
         <script src="../content_scripts/front.js"></script>
-        <script src="../content_scripts/content_scripts.js"></script>
-        <script src="../content_scripts/top.js"></script>
         <script src="default.js"></script>
+        <script src="../content_scripts/content_scripts.js"></script>
 
         <input type='file' accept='text/plain' style="display: none" />
         <div class="content">
