+++ pages/mermaid.html
@@ -7,11 +7,10 @@
         <link rel="stylesheet" href="common.css">
     </head>
     <body>
-        <script src="../content_scripts/common_content.min.js"></script>
+        <script src="../content_scripts/modules.min.js"></script>
         <script src="../content_scripts/front.js"></script>
-        <script src="../content_scripts/content_scripts.js"></script>
-        <script src="../content_scripts/top.js"></script>
         <script src="default.js"></script>
+        <script src="../content_scripts/content_scripts.js"></script>
 
         <div class="content">
             <div class="mermaid">
