+++ pages/options.html
@@ -134,11 +134,10 @@ settings.theme = `
         </div>
         <h3 id="resetSettings">Reset</h3>
         <iframe id="sandbox" allowtransparency="true" frameborder="0" scrolling="no" style="height: 0" src="/pages/sandbox.html"></iframe>
-        <script src="../content_scripts/common_content.min.js"></script>
+        <script src="../content_scripts/modules.min.js"></script>
         <script src="../content_scripts/front.js"></script>
-        <script src="../content_scripts/content_scripts.js"></script>
-        <script src="../content_scripts/top.js"></script>
         <script src="default.js"></script>
+        <script src="../content_scripts/content_scripts.js"></script>
         <script src="options.js"></script>
         <script src="ga.js"></script>
         <script async src='https://www.google-analytics.com/analytics.js'></script>
