+++ pages/popup.html
@@ -35,7 +35,7 @@ a:hover {
         <a href="/pages/pdf_viewer.html" target="_blank" class="disabledOnFirefox">PDF Viewer</a>
         <a href="/pages/markdown.html" target="_blank">Markdown preview</a>
         <a href="/pages/mermaid.html" target="_blank" class="disabledOnFirefox">Mermaid diagram generator</a>
-        <script src="../content_scripts/common_content.min.js"></script>
+        <script src="../content_scripts/modules.min.js"></script>
         <script src="popup.js"></script>
     </body>
 </html>
