+++ pages/start.html
@@ -8,11 +8,10 @@
         <link rel="stylesheet" href="start.css">
     </head>
     <body>
-        <script src="../content_scripts/common_content.min.js"></script>
+        <script src="../content_scripts/modules.min.js"></script>
         <script src="../content_scripts/front.js"></script>
-        <script src="../content_scripts/content_scripts.js"></script>
-        <script src="../content_scripts/top.js"></script>
         <script src="default.js"></script>
+        <script src="../content_scripts/content_scripts.js"></script>
 
         <script id="quickIntroSource" type="text/template">
 ## Get started with Surfingkeys.
